perf: f32 used (faster than f64)

+++ src/io.rs
@@ -2,15 +2,15 @@ extern crate mnist;
 use mnist::{Mnist, MnistBuilder};
 
 pub struct TrainData {
-	//train_array_2d: Vec<Vec<f64>>,
+	//train_array_2d: Vec<Vec<f32>>,
 	//train_label: Vec<u8>,
-	pub training_data: Vec<Vec<Vec<f64>>>,
-	//validation_array_2d: Vec<Vec<f64>>,
+	pub training_data: Vec<Vec<Vec<f32>>>,
+	//validation_array_2d: Vec<Vec<f32>>,
 	//validation_label: Vec<u8>,
-	pub validation_data: Vec<(Vec<f64>, i32)>,
-	//test_array_2d: Vec<Vec<f64>>,
+	pub validation_data: Vec<(Vec<f32>, i32)>,
+	//test_array_2d: Vec<Vec<f32>>,
 	//test_label: Vec<u8>,
-	pub test_data: Vec<(Vec<f64>, i32)>,
+	pub test_data: Vec<(Vec<f32>, i32)>,
 }
 
 pub fn import_images(training_size: usize, test_size: usize) -> TrainData {
@@ -46,27 +46,27 @@ pub fn import_images(training_size: usize, test_size: usize) -> TrainData {
 }
 
 // Designed to normalize Pixels without color-data so 255 is hardcoded
-fn normalize(x: &u8) -> f64 {
-	*x as f64 / 255.0
+fn normalize(x: &u8) -> f32 {
+	*x as f32 / 255.0
 }
 
-fn pack_images_vec(input_vector: Vec<u8>, rows: u8, cols: u8) -> Vec<Vec<f64>> {
-	let input: Vec<f64> = input_vector
+fn pack_images_vec(input_vector: Vec<u8>, rows: u8, cols: u8) -> Vec<Vec<f32>> {
+	let input: Vec<f32> = input_vector
 		.iter()
 		.map(normalize)
 		.collect();
 	
 	let image_len = (rows as i32 * cols as i32) as usize;
-	let target_vec: Vec<Vec<f64>> = input.chunks(image_len)
+	let target_vec: Vec<Vec<f32>> = input.chunks(image_len)
 		.map(|x| x.to_vec())    
 		.collect();
 	
 	target_vec
 }
 
-fn pack_training_data(data: &[Vec<f64>], label: &[u8]) -> Vec<Vec<Vec<f64>>> {
-	let mut td: Vec<Vec<Vec<f64>>> = Vec::new();
-	let num_vec: Vec<Vec<f64>> = label.iter().map(vectorise_num).collect();
+fn pack_training_data(data: &[Vec<f32>], label: &[u8]) -> Vec<Vec<Vec<f32>>> {
+	let mut td: Vec<Vec<Vec<f32>>> = Vec::new();
+	let num_vec: Vec<Vec<f32>> = label.iter().map(vectorise_num).collect();
 
 	for (x, y) in data.iter().zip(&num_vec) {
 		td.push(vec![x.clone(), y.clone()])
@@ -74,8 +74,8 @@ fn pack_training_data(data: &[Vec<f64>], label: &[u8]) -> Vec<Vec<Vec<f64>>> {
 	td
 }
 
-fn pack_vector_data(data: &[Vec<f64>], label: &Vec<u8>) -> Vec<(Vec<f64>, i32)>{
-	let mut result: Vec<(Vec<f64>, i32)> = Vec::new();
+fn pack_vector_data(data: &[Vec<f32>], label: &Vec<u8>) -> Vec<(Vec<f32>, i32)>{
+	let mut result: Vec<(Vec<f32>, i32)> = Vec::new();
 	for (x, y) in data.iter().zip(label) {
 		result.push((x.clone(), *y as i32))
 	}
@@ -84,7 +84,7 @@ fn pack_vector_data(data: &[Vec<f64>], label: &Vec<u8>) -> Vec<(Vec<f64>, i32)>{
 
 // Creates a 1d vector with size 10 (0..9) which contains a 1 a the 
 // index-location of the input @num
-pub fn vectorise_num(num: &u8) -> Vec<f64> {
+pub fn vectorise_num(num: &u8) -> Vec<f32> {
 	let mut out = vec![0.; 10];
 	out[*num as usize] = 1.;
 	out
